Add Hero3 Plus Black firmware section patching tool
Add an extremely scary tool which can replace any section
in the Hero3+ Black firmware file (HD3.11-firmware.bin)
with an arbitrary file of equal size (which will virtually
always be a modified version of that section), patching the
section and global CRCs as appropriate. The tool is very
very heavily based on the Hero4 version, with only minor
differences to account for the alternate CRC position and
encoding.

WARNING! Because this tool automatically fixes up the
firmware CRCs, it is very very easy to make a mistake and
end up with a valid-looking firmware file, which will brick
your camera when flashed.

UNLESS YOU WANT A QUICK AND EASY WAY TO BRICK YOUR CAMERA,
YOU PROBABLY DO NOT WANT TO USE THIS PROGRAM!! If you do
decide to use it, you agree to bear the sole responsibility
for the consequences of its use.

Signed-off-by: evilwombat <evilwombat@server.fake>